Update task_fn inside main and add results
task_fn notes:

Here we are using the args : [uxsched<fn_t>* sched, int runner_id]
to make use of safepoint_check as this functionality is a part of
sched and needs to be of the same schedular object. Either passing sched
as Arg or using global pointer works.

But same is not true for runner_id, this need to be an Arg to the function,
and more importantly, needs to be an Arg for the lambda function. We pass
the runner_id to the task will executing func inside the runner, which is
utilized by safepoint_check to extract the runner information.

Signed-off-by: neel <neelreddy12@gmail.com>